requirements.txt
Django==4.2
gunicorn==20.1.0
Pillow==9.5.0
django-ckeditor==6.5.1
django-jazzmin
psycopg2-binary==2.9.6